+++ src/main/kotlin/io/snyk/plugin/settings/SnykProjectSettingsConfigurable.kt
@@ -29,8 +29,7 @@ class SnykProjectSettingsConfigurable(
     private val settingsStateService
         get() = pluginSettings()
 
-    var snykSettingsDialog: SnykSettingsDialog =
-        SnykSettingsDialog(project, settingsStateService, this)
+    var snykSettingsDialog: SnykSettingsDialog = SnykSettingsDialog(project, settingsStateService, this)
 
     override fun getId(): String = "io.snyk.plugin.settings.SnykProjectSettingsConfigurable"
 
@@ -49,6 +48,8 @@ class SnykProjectSettingsConfigurable(
             snykSettingsDialog.getCliBaseDownloadURL() != settingsStateService.cliBaseDownloadURL ||
             snykSettingsDialog.isScanOnSaveEnabled() != settingsStateService.scanOnSave ||
             snykSettingsDialog.getCliReleaseChannel() != settingsStateService.cliReleaseChannel ||
+            snykSettingsDialog.getNetNewIssuesSelected() != settingsStateService.netNewIssues ||
+
             isAuthenticationMethodModified()
 
     private fun isAuthenticationMethodModified() =
@@ -117,6 +118,10 @@ class SnykProjectSettingsConfigurable(
                 handleReleaseChannelChanged()
             }
 
+            if (snykSettingsDialog.getNetNewIssuesSelected() != pluginSettings().netNewIssues) {
+                settingsStateService.netNewIssues = snykSettingsDialog.getNetNewIssuesSelected()
+            }
+
             LanguageServerWrapper.getInstance().updateConfiguration()
         }
 
@@ -136,27 +141,26 @@ class SnykProjectSettingsConfigurable(
     private fun handleReleaseChannelChanged() {
         settingsStateService.cliReleaseChannel = snykSettingsDialog.getCliReleaseChannel().trim()
         var notification: Notification? = null
-        val downloadAction =
-            object : AnAction("Download") {
-                override fun actionPerformed(e: AnActionEvent) {
-                    getSnykTaskQueueService(project)?.downloadLatestRelease(true)
-                        ?: SnykBalloonNotificationHelper.showWarn("Could not download Snyk CLI", project)
-                    notification?.expire()
-                }
+        val downloadAction = object : AnAction("Download") {
+            override fun actionPerformed(e: AnActionEvent) {
+                getSnykTaskQueueService(project)?.downloadLatestRelease(true) ?: SnykBalloonNotificationHelper.showWarn(
+                    "Could not download Snyk CLI",
+                    project
+                )
+                notification?.expire()
             }
-        val noAction =
-            object : AnAction("Cancel") {
-                override fun actionPerformed(e: AnActionEvent) {
-                    notification?.expire()
-                }
+        }
+        val noAction = object : AnAction("Cancel") {
+            override fun actionPerformed(e: AnActionEvent) {
+                notification?.expire()
             }
-        notification =
-            SnykBalloonNotificationHelper.showInfo(
-                "You changed the release channel. Would you like to download a new Snyk CLI now?",
-                project,
-                downloadAction,
-                noAction,
-            )
+        }
+        notification = SnykBalloonNotificationHelper.showInfo(
+            "You changed the release channel. Would you like to download a new Snyk CLI now?",
+            project,
+            downloadAction,
+            noAction,
+        )
     }
 
     private fun isTokenModified(): Boolean = snykSettingsDialog.getToken() != settingsStateService.token
@@ -171,6 +175,7 @@ class SnykProjectSettingsConfigurable(
         snykSettingsDialog.isIgnoreUnknownCA() != settingsStateService.ignoreUnknownCA
 
     private fun isAdditionalParametersModified(): Boolean =
-        isProjectSettingsAvailable(project) &&
-            snykSettingsDialog.getAdditionalParameters() != getSnykProjectSettingsService(project)?.additionalParameters
+        isProjectSettingsAvailable(project) && snykSettingsDialog.getAdditionalParameters() != getSnykProjectSettingsService(
+            project
+        )?.additionalParameters
 }